AI-Powered-Internship-Email-Automation-Agent-main/requirements.txt
google-genai
google-auth
google-auth-oauthlib
google-api-python-client
